+++ backend/spring-simple/src/main/java/com/todo/backend/util/TodoItemNotFoundException.java
@@ -1,7 +1,7 @@
 package com.todo.backend.util;
 
 public class TodoItemNotFoundException extends RuntimeException {
-    public TodoItemNotFoundException(Long id) {
+    public TodoItemNotFoundException(Integer id) {
         super("Could not find todo-item " + id);
     }
 }
